Used shared code between CPU and GPU formats for raw-SHA256. (No static method in header).
