+++ src/rawSHA256_ng_fmt_plug.c
@@ -41,7 +41,6 @@ john_register_one(&fmt_rawSHA256_ng);
 #include "common.h"
 #include "formats.h"
 #include "aligned.h"
-#include "rawSHA256_common.h"
 #include "memdbg.h"
 
 #if __AVX512__ || __MIC__
@@ -68,10 +67,11 @@ john_register_one(&fmt_rawSHA256_ng);
 #define BENCHMARK_LENGTH          -1
 
 #define MAXLEN                    55
+#define PLAINTEXT_LENGTH	  MAXLEN
 #define CIPHERTEXT_LENGTH         64
 #define DIGEST_SIZE               32
-#define BINARY_SIZE               32
-#define BINARY_ALIGN              4
+#include "rawSHA256_common.h"
+
 #define SALT_SIZE                 0
 #define SALT_ALIGN                1
 #define MIN_KEYS_PER_CRYPT        VWIDTH
@@ -154,48 +154,6 @@ john_register_one(&fmt_rawSHA256_ng);
     h    = vadd_epi32(tmp1, tmp2);                                    \
 }
 
-static struct fmt_tests tests[] = {
-	{"71c3f65d17745f05235570f1799d75e69795d469d9fcb83e326f82f1afa80dea", "epixoip"},
-	{HEX_TAG "71c3f65d17745f05235570f1799d75e69795d469d9fcb83e326f82f1afa80dea", "epixoip"},
-	{"25b64f637b373d33a8aa2b7579784e99a20e6b7dfea99a71af124394b8958f27", "doesthiswork"},
-	{"5e884898da28047151d0e56f8dc6292773603d0d6aabbdd62a11ef721d1542d8", "password"},
-	{"27c6794c8aa2f70f5f6dc93d3bfb25ca6de9b0752c8318614cbd4ad203bea24c", "ALLCAPS"},
-	{"04cdd6c523673bf448efe055711a9b184817d7843b0a76c2046f5398b5854152", "TestTESTt3st"},
-	{HEX_TAG "ef797c8118f02dfb649607dd5d3f8c7623048c9c063d532cc95c5ed7a898a64f", "12345678"},
-	{HEX_TAG "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855", ""},
-	{HEX_TAG "E3B0C44298FC1C149AFBF4C8996FB92427AE41E4649B934CA495991B7852B855", ""},
-	{"LcV6aBcc/53FoCJjXQMd7rBUDEpeevrK8V5jQVoJEhU", "password"},
-	{"$cisco4$LcV6aBcc/53FoCJjXQMd7rBUDEpeevrK8V5jQVoJEhU", "password"},
-	{"a49c2c9d0c006c8cb55a9a7a38822b83e0cd442614cb416af952fa50156761dc", "openwall"},
-	{"9e7d3e56996c5a06a6a378567e62f5aa7138ebb0f55c0bdaf73666bf77f73380", "mot\xf6rhead"},
-	{"1b4f0e9851971998e732078544c96b36c3d01cedf7caa332359d6f1d83567014", "test1"},
-	{"fd61a03af4f77d870fc21e05e7e80678095c92d808cfb3b5c279ee04c74aca13", "test3"},
-	{"d150eb0383c8ef7478248d7e6cf18db333e8753d05e15a8a83714b7cf63922b3", "thatsworking"},
-#ifdef DEBUG
-	{"c775e7b757ede630cd0aa1113bd102661ab38829ca52a6422ab782862f268646", "1234567890"},
-	{"6ed645ef0e1abea1bf1e4e935ff04f9e18d39812387f63cda3415b46240f0405", "12345678901234567890"},
-	{"f54e5c8f810648e7638d25eb7ed6d24b7e5999d588e88826f2aa837d2ee52ecd", "123456789012345678901234567890"},
-	{"a4ebdd541454b84cc670c9f1f5508baf67ffd3fe59b883267808781f992a0b1d", "1234567890123456789012345678901234567890"},
-	{"f58fffba129aa67ec63bf12571a42977c0b785d3b2a93cc0538557c91da2115d", "12345678901234567890123456789012345678901234567890"},
-	{"3874d5c9cc5ab726e6bbebadee22c680ce530004d4f0bb32f765d42a0a6c6dc1", "123456789012345678901234567890123456789012345678901"},
-	{"03c3a70e99ed5eeccd80f73771fcf1ece643d939d9ecc76f25544b0233f708e9", "1234567890123456789012345678901234567890123456789012345"},
-	{"0f46e4b0802fee6fed599682a16287d0397699cfd742025482c086a70979e56a", "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"}, // 31
-	{"c62e4615bd39e222572f3a1bf7c2132ea1e65b17ec805047bd6b2842c593493f", "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"}, // 32
-	{"d5e285683cd4efc02d021a5c62014694958901005d6f71e89e0989fac77e4072", "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"}, // 55
-	{"$cisco4$OsOmQzwozC4ROs/CzpczJoShdCeW9lp7k/tGrPS5Kog", "1"},
-	{"$cisco4$d7kgbEk.P6mpKdduC66fUy1BF0MImo3eyJ9uI/JbMRk", "openwall"},
-	{"$cisco4$p5BSCWNS3ivUDpZlWthR.k4Q/xWqlFyEqXdaPikHenI", "2"},
-	{"$cisco4$HwUf7ev9Fx84X2vvspULAeDbmwlg9jgm/Wk63kc3vfU", "11"},
-	{"$cisco4$bsPEUMVATKKO9yeUlJfE3OCzHlgf0s6goJpg3P1k0UU", "test"},
-	{"$cisco4$hUsuWZSE8dZERUBYNwRK8Aa8VxEGIHsuZFUCjNj2.Ac", "verylongbutweakpassword"},
-	{"$cisco4$fLUL1VG98zYDf9Q.M40nZ5blVT3M6UBex74Blw.UDCc", "thismaximumpasswordlength"},
-	{"$cisco4$Xq81UiuCj7bz9B..EX2BZumsU/d8pF5gs2NlRMW6sTk", "applesucks"},
-	{"$cisco4$O/D/cn1nawcByQoJfBxrNnUx6jjfWV.FNFx5TzmzihU", "AppleSucks"},
-#endif
-	{NULL}
-};
-
-
 static uint32_t (*saved_key)[64];
 static uint32_t *crypt_key[ 8];
 
@@ -224,25 +182,6 @@ static void done(void)
     for (i = 0; i < 8; i++)
             MEM_FREE(crypt_key[i]);
     MEM_FREE(saved_key);
-}
-
-static void *get_binary(char *ciphertext)
-{
-    static unsigned char *out;
-    int i;
-
-    if (!out)
-        out = mem_alloc_tiny(DIGEST_SIZE, MEM_ALIGN_WORD);
-
-    ciphertext += HEX_TAG_LEN;
-
-    for(i=0; i < BINARY_SIZE; i++)
-        out[i] = atoi16[ARCH_INDEX(ciphertext[i*2])] * 16 +
-                 atoi16[ARCH_INDEX(ciphertext[i*2 + 1])];
-
-    alter_endianity(out, DIGEST_SIZE);
-
-    return (void*) out;
 }
 
 static int get_hash_0(int index) { return crypt_key[0][index] & 0xf; }
@@ -480,10 +419,10 @@ struct fmt_main fmt_rawSHA256_ng = {
         init,
         done,
         fmt_default_reset,
-        prepare,
-        valid,
-        split,
-        get_binary,
+	raw_sha256_common_prepare,
+	raw_sha256_common_valid,
+	raw_sha256_common_split,
+	raw_sha256_common_binary,
         fmt_default_salt,
 #if FMT_MAIN_VERSION > 11
 		{ NULL },
